+++ cmd/uid_provider/inprov/cmd/login.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"path/filepath"
 
+	"golang.org/x/term"
 	"inspr.dev/inspr/cmd/uid_provider/client"
 	"inspr.dev/inspr/pkg/cmd"
 )
@@ -14,8 +15,10 @@ import (
 var cl client.UIDClient
 
 type loginOptionsDT struct {
-	output string
-	stdout bool
+	output   string
+	stdout   bool
+	user     string
+	password string
 }
 
 var loginOptions = loginOptionsDT{}
@@ -27,7 +30,7 @@ login is the command responsible for associating the insprctl operations
 with an account on the UID Provider on the cluster.
 `).WithExample(
 	"log in with your user and password",
-	"inprov login usr pwd",
+	"inprov login",
 ).WithFlags(
 	&cmd.Flag{
 		Name:      "output",
@@ -42,10 +45,45 @@ with an account on the UID Provider on the cluster.
 		Value:    &loginOptions.stdout,
 		DefValue: false,
 	},
-).ExactArgs(2, loginAction)
+	&cmd.Flag{
+		Name:      "user",
+		Shorthand: "u",
+		Usage:     "set the user for the login",
+		Value:     &loginOptions.user,
+		DefValue:  "",
+	},
+	&cmd.Flag{
+		Name:      "password",
+		Shorthand: "p",
+		Usage:     "set the user's password for the login",
+		Value:     &loginOptions.password,
+		DefValue:  "",
+	},
+).NoArgs(loginAction)
 
-func loginAction(c context.Context, s []string) error {
+func loginAction(c context.Context) error {
 
+	if loginOptions.user == "" {
+		if loginOptions.password != "" {
+			return fmt.Errorf("invalid user")
+		}
+		fmt.Print("Username: ")
+		fmt.Scanln(&loginOptions.user)
+	}
+	if loginOptions.password == "" {
+		fmt.Print("Password: ")
+		password, err := term.ReadPassword(0)
+		fmt.Println()
+		if err != nil {
+			return err
+		}
+		loginOptions.password = string(password)
+	}
+
+	return login(c, loginOptions.user, loginOptions.password)
+}
+
+func login(ctx context.Context, login, password string) error {
 	var err error
 	var output io.Writer
 	var outputPath string
@@ -65,21 +103,12 @@ func loginAction(c context.Context, s []string) error {
 			return err
 		}
 	}
-
-	err = login(c, s[0], s[1], output)
-
-	// no errors
-	if err == nil {
-		fmt.Fprintln(output, "logged in succesfully")
-	}
-	return err
-}
-
-func login(ctx context.Context, login, password string, output io.Writer) error {
 	token, err := cl.Login(ctx, login, password)
 	if err != nil {
 		return err
 	}
+
+	fmt.Println("Successfully logged in!")
 
 	output.Write([]byte(token))
 	return nil